# docs/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A nice kanban board with a beautiful heart ✨",
        "title": "Goroutine kanban API",
        "contact": {}
    },
    "basePath": "/",
    "paths": {
...
        "/v1/login": {
            "post": {
                "description": "Login with email and password to get a JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login a user",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.loginBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.loginResponse"
                        }
                    },
                    "400": {
                        "description": "VALIDATION_ERROR",
                        "schema": {
                            "$ref": "#/definitions/httpschema.DetailedError"
                        }
                    },
                    "401": {
                        "description": "INVALID_CREDENTIALS or USER_NOT_FOUND",
                        "schema": {
                            "$ref": "#/definitions/httpschema.DetailedError"
                        }
                    },
                    "413": {
                        "description": "PAYLOAD_TOO_LARGE",
                        "schema": {
                            "$ref": "#/definitions/httpschema.DetailedError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httpschema.Error"
                        }
                    }
                }
            }
        },
        "/v1/register": {
            "post": {
                "description": "Register a new user with email and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "Registration details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.registerBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpschema.Status"
                        }
                    },
                    "400": {
                        "description": "VALIDATION_ERROR",
                        "schema": {
                            "$ref": "#/definitions/httpschema.DetailedError"
                        }
                    },
                    "409": {
                        "description": "USER_ALREADY_EXISTS",
                        "schema": {
                            "$ref": "#/definitions/httpschema.DetailedError"
                        }
                    },
                    "413": {
                        "description": "PAYLOAD_TOO_LARGE",
                        "schema": {
                            "$ref": "#/definitions/httpschema.DetailedError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httpschema.Error"
                        }
                    }
                }
            }
        },
...
        "/v1/whoami": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get current user ID from token. This is a protected endpoint — in addition to the responses listed below, the auth middleware may return 401 with codes INVALID_AUTH_HEADER or INVALID_TOKEN.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Get current user info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.whoAmIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: INVALID_TOKEN (handler) or INVALID_AUTH_HEADER / INVALID_TOKEN (auth middleware)",
                        "schema": {
                            "$ref": "#/definitions/httpschema.DetailedError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httpschema.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.loginBody": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "secret-password"
                }
            }
        },
        "handler.loginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "jwt-token"
                }
            }
        },
...
        "handler.registerBody": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "secret-password"
                }
            }
        },
...
        "handler.whoAmIResponse": {
            "type": "object",
            "properties": {
                "uid": {
                    "type": "string",
                    "example": "018e1000-0000-7000-8000-000000000000"
                }
            }
        },
        "httpschema.Detail": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string",
                    "example": "email"
                },
                "issues": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "must be a valid email",
                        "too short"
                    ]
                }
            }
        },
        "httpschema.DetailedError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "INVALID_CREDENTIALS"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/httpschema.Detail"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Invalid login or password"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2026-03-02T15:04:05.123Z"
                }
            }
        },
        "httpschema.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "INVALID_CREDENTIALS"
                },
                "message": {
                    "type": "string",
                    "example": "Invalid login or password"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2026-03-02T15:04:05.123Z"
                }
            }
        },
        "httpschema.Status": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter JWT token with `Bearer ` prefix, e.g. `Bearer eyJhbGciOi...`",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}